requirements.txt
fastapi
uvicorn
pymupdf
python-docx
pytesseract
Pillow
opencv-python-headless
